# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracer77 API",
        "description": "API d'authentification et de géolocalisation pour Tracer77",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.77:8000/",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/positions": {
            "post": {
                "summary": "Envoyer sa position (membre connecté)",
                "operationId": "366ecfe76f754ccef928c78d1afadecb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "example": 6.3654
                                    },
                                    "lng": {
                                        "type": "number",
                                        "example": 2.4183
                                    },
                                    "vitesse": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "direction": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "satellites": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "batterie": {
                                        "type": "number",
                                        "example": 85
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Position enregistrée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/devices/positions": {
            "post": {
                "summary": "Envoyer une position depuis un boîtier ESP32",
                "operationId": "226e3ca295feae06c16cd03cca7c8b00",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identifiant_boitier",
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "identifiant_boitier": {
                                        "type": "string",
                                        "example": "TRC-0042"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "example": 6.3654
                                    },
                                    "lng": {
                                        "type": "number",
                                        "example": 2.4183
                                    },
                                    "vitesse": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "direction": {
                                        "type": "number",
                                        "example": 0
                                    },
                                    "satellites": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "batterie": {
                                        "type": "number",
                                        "example": 72
                                    },
                                    "sos": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Position enregistrée"
                    },
                    "404": {
                        "description": "Boîtier inconnu"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum ici",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}